Add Akka DNS resolver
This adds the Akka DNS resolver at all times.

We should be adding the resolver only if service locator is enabled, and enable the service locator if Akka cluster is enabled.

However, I'm having difficulty to set service locator enabled flag when Akka cluster is enabled due to service locator enabled is a SBT setting, while enable Akka cluster is a SBT task.