Avoids touching unchanged rp-application.conf on rebuilds.
Before this change, docker:publishLocal on an unchanged project would only reuse cached layers up to the ADD step in the Docker file. The ADD layer and subsequent ones had to be rebuilt from scratch because one jar changed. This ADD step creates a layer with all the jars so it is an expensive step, especially when you consider that docker push that occurs in docker:publish. The jar that changed was the main application jar, and it was rebuilt because rp-application.conf was systematically regenerated.

Now rp-application.conf is a managed resource that uses Tracked.inputChanged to avoid trigerring subsequent steps when it is unchanged.

Anecdotally, I saw this speedup on a docker:publishLocal in a multi-project build: it went from 216s to 68s. A factor of more than 3. On docker:publish, especially with slow connections, the impact could be greater.