Add support for base docker images based on deb/rpm
With this pull request, the requiredAlpinePackages and
alpinePackages settings are deprecated in favor of
the new settings userPackages and allDockerPackages,
while retaining compatibility. The new setting
packagingFormat is introduced, with a default
value of "apk", in order to use the Alpine package
manager as default. The setting can be changed to
apt, apt-get, dnf, or yum, in order to use the
package manager more suitable to a different docker
base image selected by dockerBaseImage.