Handle non-defined rpDockerPublish
In ff298c8dd8d4bac3385f5e0e0107378ad0457563 I tried to more accurately
model what I assumed was the intended behaviour, by wrapping the
invocation of rpDockerPublish in a Task.  In doing so, however, I
dropped the .?, which originally allowed for rpDockerPublish to be
undefined.  This caused a regression.

At this point I'm not sure why SbtReactiveAppPluginAll even exists.  Why
not redefine publish in Docker in SbtReactiveAppPlugin, where
rpDockerPublish is actually defined?  I went with the smaller change.