When building rp-application.conf, also consider dependent subprojects
When assembling the `rp-application.conf` file, the current code
explores the jars in the `dependencyClasspath` and the unmanaged
resources of the current subproject.
This pull request also includes in the file the unmanaged resources
of the other sbt subprojects that are dependencies of the current
subproject.

See https://github.com/lightbend/sbt-reactive-app/issues/148